+++ coast/modules/StdDataAccess/HTTPResponseMapper.cpp
@@ -6,8 +6,11 @@
  * the license that is included with this library/application in the file license.txt.
  */
 #include "HTTPResponseMapper.h"
-#include "Tracer.h"
+
 #include "HTTPConstants.h"
+#include "Tracer.h"
+
+#include <istream>
 
 RegisterResultMapper(HTTPResponseMapper);
 
@@ -18,7 +21,7 @@ namespace {
 			message.Trim(message.Length() - 1L);
 		}
 	}
-}
+}  // namespace
 
 bool HTTPResponseMapper::DoPutStream(const char *key, std::istream &is, Context &ctx, ROAnything conf) {
 	StartTrace1(HTTPResponseMapper.DoPutStream, NotNull(key));
@@ -41,5 +44,5 @@ bool HTTPResponseMapper::DoPutStream(const char *key, std::istream &is, Context 
 	}
 	TraceAny(response, "Parsed response is: ");
 	ResultMapper::DoPutAny("HTTPResponse", response, ctx, conf);
-	return (is.good() != 0);
+	return (static_cast<int>(is.good()) != 0);
 }
